+++ api/src/main/java/nl/bss/carrentapi/api/models/interfaces/CostCalculable.java
@@ -3,5 +3,5 @@ package nl.bss.carrentapi.api.models.interfaces;
 public interface CostCalculable {
     public Double calculateMonthlyCost();
 
-    public Double calculateCostForKms(Double kms);
+    public Double calculateCostForKms(long kms);
 }
